worker/requirements.txt
# Temporal
temporalio==1.7.0

# AI
anthropic==0.34.0
google-generativeai==0.8.0
openai==1.40.0
pydantic-ai==0.1.4

# Research
httpx==0.27.0
beautifulsoup4==4.12.0

# Database
psycopg[binary]==3.2.0

# Media
cloudinary==1.40.0
pillow==10.4.0
replicate==0.30.0

# Utilities
python-dotenv==1.0.0
pydantic==2.8.0
python-slugify==8.0.0
ulid-py==1.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
